+++ packages/patternfly-3/react-console/src/index.js
@@ -1,4 +1,5 @@
 export * from './SerialConsole';
+export * from './SpiceConsole';
 export * from './VncConsole';
 export * from './AccessConsoles';
 export * from './DesktopViewer';