+++ packages/patternfly-4/react-core/src/components/LoginPage/index.d.ts
@@ -1,5 +1,0 @@
-export { default as LoginForm, LoginFormProps } from './LoginForm';
-export { default as LoginPage, LoginPageProps } from './LoginPage';
-export { default as LoginFooterItem } from './LoginFooterItem';
-export { default as LoginMainFooterBandItem } from './LoginMainFooterBandItem';
-export { default as LoginMainFooterLinksItem } from './LoginMainFooterLinksItem';