+++ packages/patternfly-4/react-core/src/declarations.d.ts
@@ -1,6 +1,6 @@
 // We use a JS babel plugin to resolve these imports that `tsc` doesn't
 // know about, so pretend that everything will work
-declare module '@patternfly/patternfly/components/*.css' {
+declare module '@patternfly/patternfly/*.css' {
   const styles: any;
   export default styles;
 }